+++ example/lib/token_example.dart
@@ -8,15 +8,19 @@ void main() async {
   String baseUrl = 'http://example.com/apirest.php';
 
   // Create the client BUT doesn't get the session token
-  GlpiClient client = GlpiClient.withToken(baseUrl, userToken, appToken);
+  GlpiClient client = GlpiClient(baseUrl, appToken: appToken);
+  String? sessionToken;
 
   // Get the session token
   try {
-    await client.initSession();
+    final response = await client.initSessionUserToken(userToken);
+    sessionToken = response['session_token'];
+
+    client.sessionToken = sessionToken;
   } on GlpiException catch (e) {
     // In case of will get the http status code (e.code) and the message (e.reason['error'])
     print('${e.code} - ${e.error} - ${e.message}');
   }
 
-  print(client.sessionToken);
+  print(sessionToken);
 }